### 2024-05-31 #6821 @editorialbot
[PRE REVIEW]: plasticparcels: A python package for marine plastic dispersal simulations and parameterisation development using parcels
**Submitting author:** @michaeldenes (<a href="http://orcid.org/0000-0003-1868-0043">Michael C Denes</a>)
**Repository:** https://github.com/OceanParcels/plasticparcels
**Branch with paper.md** (empty if default branch):
**Version:** v0.2.0
