### 2024-06-03 #6821 @kthyng
* the tests directory is very light but I see there are example notebooks that potentially could be run to verify software behavior. What do you think @michaeldenes?
* The size of this package is borderline so I am going to ping the editorial board to see if it is in scope in terms of substantial scholarly effort.
