### 2024-06-04 #6821 @michaeldenes
Hi @kthyng,

We realised after your comment that the tests directory is very light indeed, I'm currently in the processes of adding a number of new unit tests to ensure the functionality works as expected.

As for potential reviewers, we suggest from the following:
Mattia Almansi - malmans2
Tomas Chor - tomchor
Pål Næverlid Sævik - pnsaevik
George Datseris - Datseris
Zhen Wu - zhenwu0728
Gael Forget - gaelforget

Thanks,
Michael
