+++ tests/test_constructors.py
@@ -1,19 +1,146 @@
+import parcels.application_kernels
 import plasticparcels as pp
 import parcels
 from datetime import datetime, timedelta
+import pytest
 
+
+def make_simple_fieldset():
+    fieldset = parcels.FieldSet.from_data({'U': 0, 'V': 0},
+                                          {'lon': 0, 'lat': 0},
+                                          mesh='spherical')
+    return fieldset
+
+
+# Test the create_hydrodynamic_fieldset() function
+def test_create_hydrodynamic_fieldset():
+    settings_file = 'tests/test_data/test_settings.json'
+    settings = pp.utils.load_settings(settings_file)
+    settings['simulation'] = {'startdate': datetime.strptime('2020-01-04-00:00:00', '%Y-%m-%d-%H:%M:%S'),
+                                'runtime': timedelta(days=1),
+                                'outputdt': timedelta(hours=1),
+                                'dt': timedelta(minutes=20),
+                                }
+
+    fieldset = pp.constructors.create_hydrodynamic_fieldset(settings)
+
+    assert isinstance(fieldset, parcels.FieldSet)
+
+# Test the create_fieldset() function
 def test_create_fieldset():
     settings_file = 'tests/test_data/test_settings.json'
     settings = pp.utils.load_settings(settings_file)
     settings = pp.utils.download_plasticparcels_dataset('NEMO0083', settings, 'input_data')
 
-    settings['simulation'] = {'startdate': datetime.strptime('2020-01-04-00:00:00', '%Y-%m-%d-%H:%M:%S'), # Start date of simulation
-                                'runtime': timedelta(days=2),        # Runtime of simulation
-                                'outputdt': timedelta(hours=1),      # Timestep of output
-                                'dt': timedelta(minutes=20),          # Timestep of advection
+    settings['simulation'] = {'startdate': datetime.strptime('2020-01-04-00:00:00', '%Y-%m-%d-%H:%M:%S'),
+                                'runtime': timedelta(days=1),
+                                'outputdt': timedelta(hours=1),
+                                'dt': timedelta(minutes=20),
                                 }
     
     fieldset = pp.constructors.create_fieldset(settings)
 
     assert isinstance(fieldset, parcels.FieldSet)
 
+# Test three different initialisation maps
+@pytest.mark.parametrize('initialisation_map', ['coastal', 'fisheries', 'rivers'])
+def test_create_particleset_from_map(initialisation_map):
+    settings_file = 'tests/test_data/test_settings.json'
+    settings = pp.utils.load_settings(settings_file)
+    settings = pp.utils.download_plasticparcels_dataset('NEMO0083', settings, 'input_data')
+
+    settings['simulation'] = {'startdate': datetime.strptime('2020-01-04-00:00:00', '%Y-%m-%d-%H:%M:%S'),
+                                'runtime': timedelta(days=1),
+                                'outputdt': timedelta(hours=1),
+                                'dt': timedelta(minutes=20),
+                                }
+ 
+    settings['plastictype'] = {'wind_coefficient' : 0.01,  # Percentage of wind to apply to particles
+                                'plastic_diameter' : 0.001, # Plastic particle diameter (m)
+                                'plastic_density' : 1030.,  # Plastic particle density (kg/m^3)
+                                }
+    
+    settings['release'] = {'initialisation_type': initialisation_map,
+                            'country': 'Italy',
+                            }
+    
+    fieldset = make_simple_fieldset()
+    pset = pp.constructors.create_particleset_from_map(fieldset, settings)
+
[…]
