### 2024-06-04 3d947ec
Adding all combinations of fields in fieldset testing
