### 2024-06-04 f7aca6e
Unit tests for all of the different kernels.
