+++ tests/test_constructors.py
@@ -11,6 +11,7 @@ def make_simple_fieldset():
                                           mesh='spherical')
     return fieldset
 
+
 def make_standard_simulation_settings():
     simulation_settings = {'startdate': datetime.strptime('2020-01-04-00:00:00', '%Y-%m-%d-%H:%M:%S'),
                                 'runtime': timedelta(days=2),
@@ -18,6 +19,7 @@ def make_standard_simulation_settings():
                                 'dt': timedelta(minutes=20),
                                 }
     return simulation_settings
+
 
 def make_standard_plastictype_settings():
     plastictype_settings = {'wind_coefficient' : 0.01,  # Percentage of wind to apply to particles
@@ -37,6 +39,7 @@ def test_create_hydrodynamic_fieldset():
 
     assert isinstance(fieldset, parcels.FieldSet)
 
+
 # Test the create_fieldset() function
 @pytest.mark.parametrize('use_3D', [True, False])
 @pytest.mark.parametrize('use_biofouling', [True, False])
@@ -49,16 +52,17 @@ def test_create_fieldset(use_3D, use_biofouling, use_stokes, use_wind, use_mixin
     settings = pp.utils.download_plasticparcels_dataset('NEMO0083', settings, 'input_data')
 
     settings['simulation'] = make_standard_simulation_settings()
-    
+
     settings['use_3D'] = use_3D
     settings['use_biofouling'] = use_biofouling
     settings['use_stokes'] = use_stokes
     settings['use_wind'] = use_wind
     settings['use_mixing'] = use_mixing
-    
+
     fieldset = pp.constructors.create_fieldset(settings)
 
     assert isinstance(fieldset, parcels.FieldSet)
+
 
 # Test three different initialisation maps
 @pytest.mark.parametrize('initialisation_map', ['coastal', 'fisheries', 'rivers'])
@@ -68,17 +72,18 @@ def test_create_particleset_from_map(initialisation_map):
     settings = pp.utils.download_plasticparcels_dataset('NEMO0083', settings, 'input_data')
 
     settings['simulation'] = make_standard_simulation_settings()
- 
+
     settings['plastictype'] = make_standard_plastictype_settings()
-    
+
     settings['release'] = {'initialisation_type': initialisation_map,
                             'country': 'Italy',
                             }
-    
+
     fieldset = make_simple_fieldset()
     pset = pp.constructors.create_particleset_from_map(fieldset, settings)
 
     assert isinstance(pset, parcels.ParticleSet)
+
 
 # Test the two global concentration release map types
 @pytest.mark.parametrize('concentration_type', ['Beach', 'Ocean'])
@@ -88,17 +93,18 @@ def test_create_particleset_from_map_concentrations(concentration_type):
     settings = pp.utils.download_plasticparcels_dataset('NEMO0083', settings, 'input_data')
 
     settings['simulation'] = make_standard_simulation_settings()
- 
+
     settings['plastictype'] = make_standard_plastictype_settings()
-    
+
     settings['release'] = {'initialisation_type': 'global_concentrations',
                             'concentration_type': concentration_type,
                             }
-    
+
     fieldset = make_simple_fieldset()
     pset = pp.constructors.create_particleset_from_map(fieldset, settings)
 
     assert isinstance(pset, parcels.ParticleSet)
+
 
 # Test three different initialisation maps
 @pytest.mark.parametrize('use_3D', [True, False])
@@ -118,7 +124,7 @@ def test_create_kernel(use_3D, use_biofouling, use_stokes, use_wind, use_mixing)
     settings['use_stokes'] = use_stokes
     settings['use_wind'] = use_wind
[…]
