+++ tests/test_kernels.py
@@ -14,6 +14,7 @@ def make_standard_simulation_settings():
                                 }
     return simulation_settings
 
+
 def make_standard_plastictype_settings():
     #Use tiny wind percentage because test data set is not large and wind speeds are quick!
     plastictype_settings = {'wind_coefficient' : 0.0001,  # Percentage of wind to apply to particles
@@ -21,6 +22,7 @@ def make_standard_plastictype_settings():
                                 'plastic_density' : 1030.,  # Plastic particle density (kg/m^3)
                                 }
     return plastictype_settings
+
 
 def make_standard_particleset(fieldset, settings):
     # Generate a particleset that has particles in the test domain
@@ -30,8 +32,9 @@ def make_standard_particleset(fieldset, settings):
     # Only keep particles in the test domain
     keep_particles = (pset.lon >17) & (pset.lon <20) & (pset.lat < 36) & (pset.lat > 34) 
     pset.remove_booleanvector(~keep_particles)
-    
+
     return pset
+
 
 @pytest.mark.parametrize('use_3D', [True, False])
 def test_advection_only(use_3D):
@@ -41,7 +44,7 @@ def test_advection_only(use_3D):
 
     settings['simulation'] = make_standard_simulation_settings()
     settings['plastictype'] = make_standard_plastictype_settings()
-    
+
     # Turn on/off 3D advection
     settings['use_3D'] = use_3D
 
@@ -67,6 +70,7 @@ def test_advection_only(use_3D):
     # Assert that the particles move from their initial location
     assert (np.sum(np.abs(pset.lon - start_lons)) > 0.) & (np.sum(np.abs(pset.lat - start_lats)) > 0.)
 
+
 def test_settling_velocity():
     settings_file = 'tests/test_data/test_settings.json'
     settings = pp.utils.load_settings(settings_file)
@@ -74,7 +78,7 @@ def test_settling_velocity():
 
     settings['simulation'] = make_standard_simulation_settings()
     settings['plastictype'] = make_standard_plastictype_settings()
-    
+
     # Turn on 3D advection
     settings['use_3D'] = True
 
@@ -85,7 +89,7 @@ def test_settling_velocity():
     settings['use_mixing'] = False
 
     fieldset = pp.constructors.create_fieldset(settings)
-    
+
     kernels = pp.constructors.create_kernel(fieldset)
 
     pset = make_standard_particleset(fieldset, settings)
@@ -99,6 +103,7 @@ def test_settling_velocity():
     # Assert that the particles move from their initial location
     assert (np.sum(np.abs(pset.lon - start_lons)) > 0.) & (np.sum(np.abs(pset.lat - start_lats)) > 0.) & (np.sum(np.abs(pset.depth - start_depths)) > 0.)
 
+
 def test_biofouling():
     settings_file = 'tests/test_data/test_settings.json'
     settings = pp.utils.load_settings(settings_file)
@@ -106,7 +111,7 @@ def test_biofouling():
 
     settings['simulation'] = make_standard_simulation_settings()
     settings['plastictype'] = make_standard_plastictype_settings()
-    
+
     # Turn on biofouling
     settings['use_biofouling'] = True
 
@@ -117,7 +122,7 @@ def test_biofouling():
     settings['use_mixing'] = False
 
     fieldset = pp.constructors.create_fieldset(settings)
-    
+
     kernels = [pp.kernels.PolyTEOS10_bsq, pp.kernels.Biofouling,
                pp.kernels.checkThroughBathymetry, pp.kernels.checkErrorThroughSurface,
                pp.kernels.deleteParticle]
@@ -131,6 +136,7 @@ def test_biofouling():
[…]
