### 2024-06-07 80e14b6
Refactored the create particleset functionality, so that users can pass their own release locations if they'd like
