### 2024-06-07 29f36bd
A fix to create_particleset to allow passing of lists (and not just numpy arrays)
