+++ plasticparcels/constructors.py
@@ -221,8 +221,8 @@ def create_particleset(fieldset, settings, release_locations):
     """
 
     # Set the longitude, latitude, and plastic amount per particle
-    lons = release_locations['lons']
-    lats = release_locations['lats']
+    lons = np.array(release_locations['lons'])
+    lats = np.array(release_locations['lats'])
     if 'plastic_amount' in release_locations.keys():
         plastic_amounts = release_locations['plastic_amount']
     else:
