### 2024-06-07 62f4938
New unit tests for the create_particleset functionality
