### 2024-06-07 0d33c5d
Changing from using an initialisation map to a list of particles in the kernels unit testing. This allows us to remove superfluous code (like the `download_plasticparcels_dataset()` in tests that don't use the unbeaching kernel.
