### 2024-06-07 b195ac3
Changing from `zero_like` to `nan_like` in the `create_particleset()` function, when users do not provide a plastic amount per particle.
