### 2024-06-07 ac570e8
Removing some references that aren't used and have no DOI
