+++ docs/paper/paper.bib
@@ -29,25 +29,6 @@
  year = {1994}
 }
 
-@misc{Allen1999,
- title = {{Review of leeway: field experiments and implementation. USCG R&D center technical report CG-D-08-99}},
- author = {Allen, A.A., Plourde, J.V.},
- year = {1999},
- month = {April},
- howpublished = {Available via: https://ntrl.ntis.gov/NTRL/},
- note = {{Accessed: 18 April 2024}}
-}
-
-
-@misc{Allen2005,
- title = {{Leeway divergence. USCG R&D center technical report CG-D-05-05}},
- author = {Allen, A.A.},
- year = {2005},
- month = {January},
- howpublished = {Available via: https://ntrl.ntis.gov/NTRL/},
- note = {{Accessed: 18 April 2024}}
-}
-
 @article{Grawe2012,
  author = {Gräwe, Ulf and Deleersnijder, Eric and Shah, Syed Hyder Ali Muttaqi and Heemink, Arnold Willem},
  doi = {10.1007/s10236-012-0523-y},
@@ -62,8 +43,6 @@
  volume = {62},
  year = {2012}
 }
-
-
 
 @article{Lebreton2012,
  author = {Lebreton, L.C.-M. and Greer, S.D. and Borrero, J.C.},
@@ -418,11 +397,4 @@
  url = {http://dx.doi.org/10.1038/s41561-023-01216-0},
  volume = {16},
  year = {2023}
-}
-
-@misc{CMS,
- title = {{Copernicus Marine Service}},
- howpublished = {https://marine.copernicus.eu/},
- note = {{Accessed: 27 March 2024}},
- year = {2024}
 }
