+++ docs/paper/paper.md
@@ -23,7 +23,7 @@ link-bibliography: false
 # Summary
 `plasticparcels` is a python package for simulating the transport and dispersion of plastics in the ocean. The tool is based on `v3.0.3` of the `parcels` computational Lagrangian ocean analysis framework [@Lange2017; @Delandmeter2019], providing a modular and customisable collection of methods, notebooks, and tutorials for advecting virtual plastic particles with a wide range of physical properties. The tool applies a collection of physical processes to the virtual particles, such as Stokes drift, wind-induced drift, biofouling, and turbulent mixing, via custom particle behaviour programmed in the form of `Kernels`. In addition to the fine-scale physics parameterisations, `plasticparcels` provides global particle initialisation maps that represent best estimates for plastic pollution emissions along coastlines [@Jambeck2015], from river sources [@Meijer2021], and in the open-ocean from fishing-related activities [@Kroodsma2018], as well as a current best estimate of buoyant plastic concentrations globally [@Kaandorp2023]. We envisage `plasticparcels` as a tool for easy-to-run plastic dispersal simulations; as well as for rapid prototyping, development, and testing of new fine-scale physics parameterisations.
 
-The current version supports nano- and microplastic behaviour, with support for macroplastics planned in the near-future. It has been designed for use with hydrodynamic and biogeochemical data from the Copernicus Marine Service [@CMS], providing new plastic modelling capabilities as part of the NECCTON project. `plasticparcels` is easily adapted to run on local machines and high-performance computing (HPC) architecture with various hydrodynamic, biogeochemical, and other model fields as input. A future goal is to embed `plasticparcels` within a cloud platform to allow for even more rapid prototyping, development, and simulations.
+The current version supports nano- and microplastic behaviour, with support for macroplastics planned in the near-future. It has been designed for use with hydrodynamic and biogeochemical data from the [Copernicus Marine Service](https://marine.copernicus.eu/), providing new plastic modelling capabilities as part of the NECCTON project. `plasticparcels` is easily adapted to run on local machines and high-performance computing (HPC) architecture with various hydrodynamic, biogeochemical, and other model fields as input. A future goal is to embed `plasticparcels` within a cloud platform to allow for even more rapid prototyping, development, and simulations.
 
 
 # Statement of need
