### 2024-06-07 ec00498
Fixing minor issue if create_particleset test
