### 2024-06-07 8326777
nan_like isn't a function, and from old bug this didn't get run. Now using full_like, filling with np.nan for cases where particle amounts are not provided.
