### 2024-06-14 #6821 @kthyng
Ok this has passed the scope query, and is added to the waitlist to wait for an appropriate editor. How long do you expect the tests development to take?
