### 2024-06-14 #6821 @michaeldenes
Thanks @kthyng, I've created a number of unit tests, now released in the main branch.

We can create a new release of the code if that is useful?
I might include a couple of enhancements in the meantime if that's the case!
