### 2024-06-20 #6821 @kthyng
Hi @michaeldenes! You don't need to create a new release but you can, the reviewers will review whatever the newest code available is when they start. If you keep significantly developing the code at the point they are reviewing and it's unrelated to their work/comments/questions and it's new stuff, then it would be good to clarify that with separate versions or have that work in a branch (keeping in mind sometimes the reviews can take awhile depending on how they go). Does that make sense?
