### 2024-06-20 #6821 @michaeldenes
Thanks @kthyng, I'll keep things separated in versions/branches, it would only be a few minor optimisations (to reduce file sizes, etc.), but nothing game-changing. Cheers!
