### 2024-08-09 #6821 @AnjaliSandip
> I would be happy to if I have a reasonable amount of time. When would you need my review? On Aug 8, 2024, at 1:55 PM, Anjali Sandip ***@***.***> wrote: ﻿ CAUTION: This email originated from outside the organization. DO NOT CLICK ON LINKS or OPEN ATTACHMENTS unless you know and trust the sender. ***@***.***<https://urldefense.com/v3/__https://github.com/selipot__;!!KVu0SnhVq1hAFvslES2Y!KHCVRKlRn6ZaO6m24OaNHzfS9MNIjzI1rjnHDjk10gtBcMJ6msbozyWd4rkvgqKeUr6ICRZpsTh8yeNyDRbhWDCE$>, @philippemiron<https://urldefense.com/v3/__https://github.com/philippemiron__;!!KVu0SnhVq1hAFvslES2Y!KHCVRKlRn6ZaO6m24OaNHzfS9MNIjzI1rjnHDjk10gtBcMJ6msbozyWd4rkvgqKeUr6ICRZpsTh8yeNyDVA8VoGC$>, and @travis-j-hahn<https://urldefense.com/v3/__https://github.com/travis-j-hahn__;!!KVu0SnhVq1hAFvslES2Y!KHCVRKlRn6ZaO6m24OaNHzfS9MNIjzI1rjnHDjk10gtBcMJ6msbozyWd4rkvgqKeUr6ICRZpsTh8yeNyDUk1Fsoi$>, would you review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html<https://urldefense.com/v3/__https://joss.readthedocs.io/en/latest/review_criteria.html__;!!KVu0SnhVq1hAFvslES2Y!KHCVRKlRn6ZaO6m24OaNHzfS9MNIjzI1rjnHDjk10gtBcMJ6msbozyWd4rkvgqKeUr6ICRZpsTh8yeNyDW8yFeJ9$>. — Reply to this email directly, view it on GitHub<https://urldefense.com/v3/__https://github.com/openjournals/joss-reviews/issues/6821*issuecomment-2276366833__;Iw!!KVu0SnhVq1hAFvslES2Y!KHCVRKlRn6ZaO6m24OaNHzfS9MNIjzI1rjnHDjk10gtBcMJ6msbozyWd4rkvgqKeUr6ICRZpsTh8yeNyDZh1z-3P$>, or unsubscribe<https://urldefense.com/v3/__https://github.com/notifications/unsubscribe-auth/ABZZWGW2MR4TY7WIID3CTLTZQOWJLAVCNFSM6AAAAABITQ2O3KVHI2DSMVQWIX3LMV43OSLTON2WKQ3PNVWWK3TUHMZDENZWGM3DMOBTGM__;!!KVu0SnhVq1hAFvslES2Y!KHCVRKlRn6ZaO6m24OaNHzfS9MNIjzI1rjnHDjk10gtBcMJ6msbozyWd4rkvgqKeUr6ICRZpsTh8yeNyDcUDm5jX$>. You are receiving this because you were mentioned.Message ID: ***@***.***>

A JOSS review is typically completed in 6 weeks.
