### 2024-08-13 OceanParcels/plasticparcels#46 @zhenwu0728
Missing API function index in documentation & missing community guidelines
As requested by [JOSS](https://github.com/openjournals/joss-reviews/issues/7094), API function index and community guidelines are missing.
