+++ plasticparcels/__init__.py
@@ -1,3 +1,5 @@
 from plasticparcels.constructors import *  # noqa
 from plasticparcels.particlefile import *  # noqa
 from plasticparcels.utils import *  # noqa
+
+from plasticparcels._version_setup import __version__, __version_tuple__  # noqa
