### 2024-08-14 c1b851e
Build API documentation
Also bugfix broken github link resolver
