+++ docs/reference.rst
@@ -1,0 +1,33 @@
+API
+===
+
+plasticparcels.constructors
+---------------------------
+
+.. automodule:: plasticparcels.constructors
+    :members:
+    :show-inheritance: yes
+
+
+plasticparcels.kernels
+---------------------------
+
+.. automodule:: plasticparcels.kernels
+    :members:
+    :show-inheritance: yes
+
+
+plasticparcels.particlefile
+---------------------------
+
+.. automodule:: plasticparcels.particlefile
+    :members:
+    :show-inheritance: yes
+
+
+plasticparcels.utils
+---------------------------
+
+.. automodule:: plasticparcels.utils
+    :members:
+    :show-inheritance: yes
