+++ docs/environment_docs.yml
@@ -27,7 +27,7 @@ dependencies:
   - numpydoc
   - nbsphinx
   - sphinx<6
-  - pandoc>1.12.1,<3
+  - pandoc>3,<4
   - pydata-sphinx-theme
   - sphinx-autobuild
   - myst-parser
