### 2024-08-15 cc091cb
Configure Ruff, add pre-commit, remove flake isort and pydocstyle
Ruff replaces flake isort and pydocstyle
