+++ tests/pyproject.toml
@@ -1,0 +1,5 @@
+[tool.ruff]
+extend = "../pyproject.toml"
+
+[tool.ruff.lint]
+ignore = ["D", "DOC"]
