### 2024-08-16 5da39de
Updating readme to include community guidelines
