+++ plasticparcels/constructors.py
@@ -10,17 +10,19 @@ from plasticparcels.utils import select_files
 
 
 def create_hydrodynamic_fieldset(settings):
-    """A constructor method to create a Parcels.Fieldset from hydrodynamic model data
+    """A constructor method to create a `Parcels.Fieldset` from hydrodynamic
+    model data.
 
     Parameters
     ----------
     settings :
-        A dictionary of settings used to create the fieldset
+        A dictionary of settings that contains an ocean model directory, a filename style,
+        and the location of the ocean model mesh file, used to create the fieldset.
 
     Returns
     -------
     fieldset
-        A parcels.FieldSet object
+        A `parcels.FieldSet` object.
     """
     # Location of hydrodynamic data
     dirread_model = os.path.join(settings['ocean']['directory'], settings['ocean']['filename_style'])
@@ -86,17 +88,21 @@ def create_hydrodynamic_fieldset(settings):
 
 
 def create_fieldset(settings):
-    """A constructor method to create a Parcels.Fieldset with all fields necessary for a plasticparcels simulation
+    """A constructor method to create a `Parcels.Fieldset` with all fields
+    necessary for a plasticparcels simulation (e.g., a hydrodynamic model
+    velocity field, biogeochemical model variable fields, a wind field, etc.).
 
     Parameters
     ----------
     settings :
-        A dictionary of model settings used to create the fieldset
+        A dictionary of settings that contains an ocean model information,
+        biogeochemical model information, wind model information,
+        and other optional settings.
 
     Returns
     -------
     fieldset
-        A parcels.FieldSet object
+        A `parcels.FieldSet` object.
     """
     # First create the hydrodynamic fieldset
     fieldset = create_hydrodynamic_fieldset(settings)
@@ -201,21 +207,22 @@ def create_fieldset(settings):
 
 
 def create_particleset(fieldset, settings, release_locations):
-    """A constructor method to create a Parcels.ParticleSet for a plasticparcels simulation
+    """A constructor method to create a `Parcels.ParticleSet` for a
+    `plasticparcels` simulation.
 
     Parameters
     ----------
     fieldset :
-        A Parcels.FieldSet object
+        A `Parcels.FieldSet` object.
     settings :
-        A dictionary of model settings, simulation settings, and release settings and plastic-type settings
+        A dictionary containing the plastic-type settings.
     release_locations :
-        A dictionary of release locations for particles
+        A dictionary containing release locations for particles.
 
     Returns
     -------
     particleset
-        A parcels.ParticleSet object
+        A parcels.ParticleSet object.
     """
     # Set the longitude, latitude, and plastic amount per particle
     lons = np.array(release_locations['lons'])
@@ -230,18 +237,6 @@ def create_particleset(fieldset, settings, release_locations):
     plastic_diameters = np.full(lons.shape, settings['plastictype']['plastic_diameter'])
     wind_coefficients = np.full(lons.shape, settings['plastictype']['wind_coefficient'])
 
-    # Create a PlasticParticle class
-    # TODO: Update to the new add_variables() approach for 3.0.2
-    # PlasticParticle = JITParticle.add_variables([
-    #     Variable('plastic_diameter', dtype=np.float32, initial=np.nan, to_write=False),
-    #     Variable('plastic_density', dtype=np.float32, initial=np.nan, to_write=False),
-    #     Variable('wind_coefficient', dtype=np.float32, initial=0., to_write=False),
-    #     Variable('settling_velocity', dtype=np.float64, initial=0., to_write=False),
-    #     Variable('seawater_density', dtype=np.float32, initial=np.nan, to_write=False),
[…]
