### 2024-08-16 a52a9fe
Updates to docstrings, stokes, wind, settling.
