+++ plasticparcels/kernels.py
@@ -8,27 +8,32 @@ import math
 
 
 def StokesDrift(particle, fieldset, time):
-    """Stokes drift kernel
+    """Stokes drift kernel.
 
     Description
     ----------
-    Using the approach in [1] assuming a Phillips wave spectrum to determine the depth dependent Stokes drift.
-    Specifically, the 'Stokes drift velocity' (u_s) is computed as per Eq. (19) in [1]
+    Using the approach in [1] assuming a Phillips wave spectrum to determine
+    the depth dependent Stokes drift. Specifically, the 'Stokes drift velocity'
+    $u_s$ is computed as per Eq. (19) in [1].
 
     We treat the Stokes drift as a linear addition to the velocity field
-        u(x,t) = u_c(x,t) + C_s * u_s(x,t)
-    where u_c is the current velocity, u_s is the Stokes drift velocity, and C_s is the depth-varying decay factor.
+        $u(x,t) = u_c(x,t) + C_s * u_s(x,t)$
+    where $u_c$ is the current velocity, $u_s$ is the Stokes drift velocity,
+    and $C_s$ is the depth-varying decay factor.
+
+    For further description, see https://plastic.oceanparcels.org/en/latest/physicskernels.html#stokes-drift
 
     Parameter Requirements
     ----------
     fieldset :
-        - Stokes drift field (U,V)  [m s-1]
-        - Peak wave period (T_p)    [s]
+        - `fieldset.Stokes_U` and `fieldset.Stokes_V`, the Stokes drift
+        velocity field. Units [m s-1]
+        - `fieldset.wave_Tp`, the peak wave period field ($T_p$). Units [s].
 
     Kernel Requirements
     ----------
     Order of Operations:
-        None
+        None - can be applied at any time.
 
     References
     ----------
@@ -67,32 +72,37 @@ def StokesDrift(particle, fieldset, time):
 
 # Wind related kernels
 def WindageDrift(particle, fieldset, time):
-    """Leeway windage kernel
+    """Leeway windage kernel.
 
     Description
     ----------
-    A simple windage kernel that applies a linear 'wind velocity' to the particle.
+    A simple windage kernel that applies a linear relative 'wind velocity'
+    to the particle.
 
     We treat the windage drift as a linear addition to the velocity field
-        u(x,t) = u_c(x,t) + C_w * u_w(x,t)
-    where u_c is the current velocity, u_w is the wind velocity at 10m height, and C_w is the windage coefficient (usually taken to be between 0.01 and 0.05,
-    depending on particle size)
+        $u(x,t) = u_c(x,t) + C_w * (u_w(x,t)-u_c(x,t))$
+    where $u_c$ is the ocean current velocity, $u_w$ is the wind velocity
+    at 10m height, and $C_w$ is the windage coefficient (usually taken to
+    be in [1%,5%], depending on particle size)
 
-    See https://doi.org/10.1007/s10652-016-9499-3 for impact of windage on LCS detection
-    or https://doi.org/10.1016/j.apor.2011.01.005 for Breivik et al 2011
+    
+    For further description, see https://plastic.oceanparcels.org/en/latest/physicskernels.html#wind-induced-drift-leeway
 
     Parameter Requirements
     ----------
     particle :
-        - wind_coefficient - the particle windage coefficient (usually taken to be between 0.01 and 0.05, depending on particle size)
+        - wind_coefficient - the particle windage coefficient in decimals
+        (usually taken to be between 0.01 and 0.05,
+        depending on particle size).
     fieldset :
-        - Wind field (U,V) at 10m height above sea surface [m s-1]
+        - `fieldset.Wind_U` and `fieldset.Wind_V`, the wind velocity field at
+        10m height above sea surface. Units [m s-1].
 
 
     Kernel Requirements
     ----------
[…]
