### 2024-08-16 817e236
Updates to biofouling kernel docstrings
