+++ plasticparcels/kernels.py
@@ -39,7 +39,6 @@ def StokesDrift(particle, fieldset, time):
     ----------
     [1] Breivik (2016) - https://doi.org/10.1016/j.ocemod.2016.01.005
 
-
     """
     # Sample the U / V components of Stokes drift
     stokes_U = fieldset.Stokes_U[time, particle.depth, particle.lat, particle.lon]
@@ -215,32 +214,33 @@ def Biofouling(particle, fieldset, time):
 
     Description
     ----------
-    Kernel to compute the settling velocity of particles due to changes in ambient algal concentrations, growth and death of attached algae based on [2].
-    The settling velocity of the particle is computed as per [1], however the particle size and density is affected by a biofouling process. The algae attached
-    to the particle (A) has a growth rate described by
-        dA/dt = C + G - M - R
+    Kernel to compute the settling velocity of particles due to changes in ambient algal
+    concentrations, growth and death of attached algae based on [2]. The settling velocity
+    of the particle is computed as per [1], however the particle size and density is
+    affected by a biofouling process. The algae attached to the particle $A$ has a growth
+    rate described by
+        $dA/dt = C + G - M - R$
 
-    Here, C models fouling of the plastic through collision with algae
-        C = beta_A * A_A / theta_pl
-            beta_A   -> Encounter rate
-            A_A      -> Ambient algal concentration
-            theta_pl -> Surface area of plastic particle
+    Here, $C$ models fouling of the plastic through collision with algae
+        $C = beta_A * A_A / theta_pl$,
+    where $beta_A$ is the encounter rate, $A_A$ is the ambient algal concentration, and
+    $theta_pl$ is the surface area of plastic particle.
 
-    G models the growth of the algae attached to the surface of the particle
-        G = mu_A * A
-            mu_A -> Algal growth rate
-            A    -> Amount of algae attached to a particle
+    $G$ models the growth of the algae attached to the surface of the particle
+        $G = mu_A * A$,
+    where $mu_A$ is the algal growth rate.
 
-    M models the grazing mortality of the algae attached to the surface of the particle
-        M = m_A * A
-            m_A -> Algal mortality rate
-            A   -> Amount of algae attached to a particle
+    $M$ models the grazing mortality of the algae attached to the surface of the particle
+        $M = m_A * A$,
+    where $m_A$ is the algal mortality rate.
 
-    R models the respiration  of the algae attached to the surface of the particle
-        R = Q_10
+    $R$ models the respiration of the algae attached to the surface of the particle
+        $R = Q_{10}^{(T-20)/10}R_{20}A$,
+    where $Q_{10}$ is the temperature coefficient, which indicates how much the respiration
+    increases when the temperature increases by 10C. $T$ is the temperature, and $R_{20}$
+    is the respiration rate.
 
-
-    TODO:Calculation steps:
+    Calculation steps:
         1. Compute the seawater dynamic viscosity from Eq. (27) in [2]
         2. Compute the kinematic viscosity from Eq. (25) in [2]
         3. Compute the dimensionless particle diameter from Eq. (4) in [2], equivalent to Eq. (6) in [1]
@@ -248,31 +248,27 @@ def Biofouling(particle, fieldset, time):
         5. Compute the settling velocity of the particle from Eq. (2) in [2], equivalent to Eq. (9) in [1]
 
 
-
-
     Parameter Requirements
     ----------
     particle :
-        - Particle diameter 'diameter'
-        - Particle density 'density'
[…]
