### 2024-08-16 f34dae6
Some updates to docstrings, still to complete
