### 2024-08-26 06b2601
Update docstrings in utils file
