### 2024-08-26 b1b6641
Cleaning up the docstrings and reformatting the mathematics
