### 2024-08-26 a94c404
FIxing up docstring math
