+++ plasticparcels/kernels.py
@@ -14,12 +14,12 @@ def StokesDrift(particle, fieldset, time):
     ----------
     Using the approach in [1] assuming a Phillips wave spectrum to determine
     the depth dependent Stokes drift. Specifically, the 'Stokes drift velocity'
-    `u_s` is computed as per Eq. (19) in [1].
+    :math:`u_s` is computed as per Eq. (19) in [1].
 
     We treat the Stokes drift as a linear addition to the velocity field
-        `u(x,t) = u_c(x,t) + C_s * u_s(x,t)`
-    where `u_c` is the current velocity, `u_s` is the Stokes drift velocity,
-    and `C_s` is the depth-varying decay factor.
+        :math:`u(x,t) = u_c(x,t) + C_s * u_s(x,t)`
+    where :math:`u_c` is the current velocity, :math:`u_s` is the Stokes drift velocity,
+    and :math:`C_s` is the depth-varying decay factor.
 
     For further description, see https://plastic.oceanparcels.org/en/latest/physicskernels.html#stokes-drift
 
@@ -28,7 +28,7 @@ def StokesDrift(particle, fieldset, time):
     fieldset :
         - `fieldset.Stokes_U` and `fieldset.Stokes_V`, the Stokes drift
         velocity field. Units [m s-1]
-        - `fieldset.wave_Tp`, the peak wave period field (`T_p`). Units [s].
+        - `fieldset.wave_Tp`, the peak wave period field (:math:`T_p`). Units [s].
 
     Kernel Requirements
     ----------
@@ -79,9 +79,9 @@ def WindageDrift(particle, fieldset, time):
     to the particle.
 
     We treat the windage drift as a linear addition to the velocity field
-        `u(x,t) = u_c(x,t) + C_w * (u_w(x,t)-u_c(x,t))`
-    where `u_c` is the ocean current velocity, `u_w` is the wind velocity
-    at 10m height, and `C_w` is the windage coefficient (usually taken to
+        :math:`u(x,t) = u_c(x,t) + C_w * (u_w(x,t)-u_c(x,t))`
+    where :math:`u_c` is the ocean current velocity, :math:`u_w` is the wind velocity
+    at 10m height, and :math:`C_w` is the windage coefficient (usually taken to
     be in [1%,5%], depending on particle size)
 
 
@@ -217,27 +217,27 @@ def Biofouling(particle, fieldset, time):
     Kernel to compute the settling velocity of particles due to changes in ambient algal
     concentrations, growth and death of attached algae based on [2]. The settling velocity
     of the particle is computed as per [1], however the particle size and density is
-    affected by a biofouling process. The algae attached to the particle `A` has a growth
+    affected by a biofouling process. The algae attached to the particle :math:`A` has a growth
     rate described by
-        `dA/dt = C + G - M - R`
+        :math:`dA/dt = C + G - M - R`
 
-    Here, `C` models fouling of the plastic through collision with algae
-        `C = beta_A * A_A / theta_pl`,
-    where `beta_A` is the encounter rate, `A_A` is the ambient algal concentration, and
-    `theta_pl` is the surface area of plastic particle.
+    Here, :math:`C` models fouling of the plastic through collision with algae
+        :math:`C = beta_A * A_A / theta_pl`,
+    where :math:`beta_A` is the encounter rate, :math:`A_A` is the ambient algal concentration, and
+    :math:`theta_pl` is the surface area of plastic particle.
 
-    `G` models the growth of the algae attached to the surface of the particle
-        `G = mu_A * A`,
-    where `mu_A` is the algal growth rate.
+    :math:`G` models the growth of the algae attached to the surface of the particle
+        :math:`G = mu_A * A`,
+    where :math:`mu_A` is the algal growth rate.
 
[…]
