### 2024-08-26 b7052df
More ` for ', and fixing biofouling math docstring to read correctly in readthedocs
