### 2024-08-26 dcda85c
Cleaning up the docstrings
