### 2024-08-26 3ba8fe6
Formatting fixes for docstrings
