### 2024-08-26 6b52cc3
Adding community guidelines text to main documentation page
