### 2024-08-26 9dc75fe
Fixes to link formatting on index page
