+++ docs/paper/paper.md
@@ -31,7 +31,7 @@ Marine plastic debris can be found almost everywhere in the ocean. A recent stud
 
 Due to its durable, inert, and cheap-to-manufacture nature, plastic has become one of the most abundant manufactured synthetic materials on Earth. Between 1950 and 2017 an estimated 8,300 million tonnes [@Geyer2017] of virgin plastic was produced, with the rate of production only set to increase. Its durability is of primary concern to the marine environment, where, without intervention, they will likely degrade and fragment into smaller pieces that will disperse across ever larger distances. These plastics interact and interfere with marine wildlife, either entangling, or being inadvertently ingested, with documented cases affecting over 900 marine species so far [@Kuhn2020]. To better understand and predict the effects of plastic pollution on the marine environment, it is of paramount importance to map where and how plastic enters our ocean, and the pathways of transport, dispersal patterns, and ultimate fate of these plastics.
 
-Lagrangian ocean analysis, where virtual particles are tracked in hydrodynamic flow fields, is widely used to uncover and investigate the pathways and timescales of the dispersion of plastic particulates in the ocean [@Lebreton2012; @Hardesty2017; @JalonRojas2019; @Chassignet2021; @Kaandorp2023]. However, two important questions arise when performing such Lagrangian simulations. Firstly, what physical processes drive the transport and dispersal of a plastic particle? The properties of plastic particles (e.g., size, shape, and density) determine what the dominant physical processes are at play, and due to the chaotic nature of the ocean, the dispersal patterns and transport behaviours of plastics will critically depend on their properties. Current state-of-the-art ocean models are either too coarse in resolution to capture these processes, or disregard these processes entirely, and so parameterising these processes is important to model and simulate their effects. Secondly, what are the initial release locations and concentrations of marine plastic pollution? Forecasting near-future spatial maps of plastic concentrations is largely an initial value problem, relying on accurate initial conditions for a realistic simulation output.
+Lagrangian ocean analysis, where virtual particles are tracked in hydrodynamic flow fields, is widely used to uncover and investigate the pathways and timescales of the dispersion of plastic particulates in the ocean [@Lebreton2012; @Hardesty2017; @JalonRojas2019; @Chassignet2021; @Kaandorp2023]. However, two important questions arise when performing such Lagrangian simulations. Firstly, what physical processes drive the transport and dispersal of a plastic particle? The properties of plastic particles (e.g., size, shape, and density) determine what the dominant physical processes are at play, and due to the chaotic nature of the ocean, the dispersal patterns and transport behaviours of plastics will critically depend on their properties. Current state-of-the-art ocean models are either too coarse in resolution to capture these processes, or disregard these processes entirely, and so parameterising these processes is important to model and simulate their effects. Secondly, what are the initial release locations and concentrations of marine plastic pollution? Forecasting near-future spatial maps of plastic concentrations is largely an initial value problem, relying on accurate initial conditions for a realistic simulation output. As yet, there is no single comprehensive source of data for estimates of current marine plastic pollution concentrations, mismanaged plastic waste along coastlines, mismanaged plastic waste in rivers, and plastic entering the ocean from fishing-related activities. Combining best estimates of current marine plastic pollution concentrations and likely future plastic pollution sources is necessary for predicting near-future plastic concentration maps in the ocean.
 
[…]
