### 2024-08-26 4a5f6d2
Update docs/paper/paper.md
Good suggestion!

Co-authored-by: Erik van Sebille <e.vansebille@uu.nl>
