### 2024-08-27 84e2d0a
Minor update to docstring
