### 2024-08-27 55ace59
Ignore D100 warning for linting.
