### 2024-08-27 7e419d3
Update API structure and Ruff lint ignore
