+++ docs/reference/particlefile.rst
@@ -1,0 +1,6 @@
+plasticparcels.particlefile
+---------------------------
+
+.. automodule:: plasticparcels.particlefile
+    :members:
+    :show-inheritance: yes
