### 2024-08-27 OceanParcels/plasticparcels#46 @michaeldenes
Thanks @zhenwu0728, we've now added these to documentation (and the community guidelines are also on the github readme page).
