### 2024-08-28 #7094 @AnjaliSandip
> Overall, the paper is clear and well written. Just 2 comments for the paper:
>
>     * Lines 59-62: the second important question needs a bit more elaboration.
>
>     * I feel like that the third section (Description of the software) could be divided and put into Summary and Statement of Need. For example, the first paragraph (Lines 79-91) goes to Statement of Need (Lines 63-73) and the second and third paragraphs (Lines 92-105) go to Summary.

@michaeldenes, Could you address the issues listed above?
