### 2024-08-30 #7094 @philippemiron
The paper is well-written and easy to follow. Also shows an extensive background and statement of need.

I only have a few comments:
- l.24: NECCTON, is it possible to include the full name of the project
- l.78: I understand what you mean but I think the description a "layer over parcels" is not very clear. Could it be renamed to something like "a parcel extension" ?
- l.79–80: Could you include a quick description of the three objects enumerated: `FieldSet`, `ParticleSet`, and `Kernels`, for people who are not fully familiar with parcels?
- Usage example, I understand this is a "part" of the example, but you could include:
`from datetime import datetime, timedelta` since you are using it below.
