### 2024-09-03 #7094 @michaeldenes
Hi @zhenwu0728 and @philippemiron, thank you both for your comments, we are currently working on implementing the required changes to the paper in this pull-request: https://github.com/OceanParcels/plasticparcels/pull/48

@zhenwu0728, on your comment, I'm not quite sure I understand your suggestion for which part should go where. From the 'Description of the software', I can see how the first paragraph (lines 78 - 90) may go at the end of the 'Statement of Need' (with a bit of rewording in the last paragraph of that section and combining it with the text from lines 78-90). For the second and third paragraphs (lines 91-104), these can go directly to the end of the 'Summary'. Was that what you had in mind?

Thanks!
