### 2024-09-04 c10e049
Final changes which fix the comments from https://github.com/openjournals/joss-reviews/issues/7094#issuecomment-2326733112
