+++ docs/paper/paper.md
@@ -25,6 +25,11 @@ link-bibliography: false
 
 The current version supports nano- and microplastic behaviour, with support for macroplastics planned in the near-future. It has been designed for use with hydrodynamic and biogeochemical data from the [Copernicus Marine Service](https://marine.copernicus.eu/), providing new plastic modelling capabilities as part of the [NECCTON (New Copernicus Capability for Trophic Ocean Networks)](https://neccton.eu/) project. `plasticparcels` is easily adapted to run on local machines and high-performance computing (HPC) architecture with various hydrodynamic, biogeochemical, and other model fields as input. A future goal is to embed `plasticparcels` within a cloud platform to allow for even more rapid prototyping, development, and simulations.
 
+The core features of `plasticparcels` are: 1) a user-friendly python notebook layer on top of `parcels` that provides a streamlined workflow for performing plastic dispersal simulations, 2) custom `parcels` kernels designed to simulate the fine-scale physical processes that influence the transport of nano- and microplastic particulates, and 3) global particle initialisation maps which represent the best estimate locations of plastic pollution emissions from coastal sources, river sources, open ocean fishing-related activity emission sources, and a current best estimate of buoyant plastic concentrations. We visualise these initialisation maps in Fig. \ref{fig:initialisation_maps}.
+
+![Particle initialisations maps based on a) coastal mismanaged plastic waste emissions [@Jambeck2015], b) riverine mismanaged plastic waste emissions [@Meijer2021], c) fishing activity related plastic emissions [@Kroodsma2018], and d) a current global surface concentration estimate [@Kaandorp2023].\label{fig:initialisation_maps}](initialisation_maps.png){width=100%}
+
+In addition, due to the flexibility of the package, users may use the functions and modular design of `plasticparcels` to enhance their existing `parcels` simulations and workflow. For example, users can use the initialisation maps, associated `ParticleSet` creation methods, and/or the custom physics kernels with their own `parcels` simulations. Post-processing and analysis of the generated trajectory datasets is purposefully left to the user, however some tutorials are provided in the [`plasticparcels` documentation](https://plastic.oceanparcels.org/en/latest/examples.html), along with the tutorials in the [`parcels` documentation](https://docs.oceanparcels.org/en/latest/documentation.html). Below we provide an example use case of `plasticparcels`.
 
 # Statement of need
 Marine plastic debris can be found almost everywhere in the ocean. A recent study estimates that there is approximately 3,200 kilotonnes of (initially) positively buoyant plastics in the global ocean in the year 2020 [@Kaandorp2023], where 59-62\% of these plastics are found at the ocean surface, 36-39\% within the deeper ocean, and 1.5-1.9\% along the coastline. They estimate that 500 kilotonnes of positively buoyant plastic enters the ocean each year, where 39-42\% originate from mismanaged waste along coastlines, 45-48\% originate from fishing-related activities (e.g. fishing lines, nets, traps, and crates), and 12-13\% from mismanaged waste entering the ocean via rivers.
@@ -35,19 +40,9 @@ Lagrangian ocean analysis, where virtual particles are tracked in hydrodynamic f
 
[…]
