+++ docs/paper/paper.md
@@ -29,7 +29,7 @@ The core features of `plasticparcels` are: 1) a user-friendly python notebook la
 
 ![Particle initialisations maps based on a) coastal mismanaged plastic waste emissions [@Jambeck2015], b) riverine mismanaged plastic waste emissions [@Meijer2021], c) fishing activity related plastic emissions [@Kroodsma2018], and d) a current global surface concentration estimate [@Kaandorp2023].\label{fig:initialisation_maps}](initialisation_maps.png){width=100%}
 
-In addition, due to the flexibility of the package, users may use the functions and modular design of `plasticparcels` to enhance their existing `parcels` simulations and workflow. For example, users can use the initialisation maps, associated `ParticleSet` creation methods, and/or the custom physics kernels with their own `parcels` simulations. Post-processing and analysis of the generated trajectory datasets is purposefully left to the user, however some tutorials are provided in the [`plasticparcels` documentation](https://plastic.oceanparcels.org/en/latest/examples.html), along with the tutorials in the [`parcels` documentation](https://docs.oceanparcels.org/en/latest/documentation.html). Below we provide an example use case of `plasticparcels`.
+In addition, due to the flexibility of the package, users may use the functions and modular design of `plasticparcels` to enhance their existing `parcels` simulations and workflow. For example, users can use the initialisation maps, associated `ParticleSet` creation methods, and/or the custom physics kernels with their own `parcels` simulations. Post-processing and analysis of the generated trajectory datasets is purposefully left to the user, however some tutorials are provided in the [`plasticparcels` documentation](https://plastic.oceanparcels.org/en/latest/examples.html), along with the tutorials in the [`parcels` documentation](https://docs.oceanparcels.org/en/latest/documentation.html). We provide an example use case of `plasticparcels` below.
 
 # Statement of need
 Marine plastic debris can be found almost everywhere in the ocean. A recent study estimates that there is approximately 3,200 kilotonnes of (initially) positively buoyant plastics in the global ocean in the year 2020 [@Kaandorp2023], where 59-62\% of these plastics are found at the ocean surface, 36-39\% within the deeper ocean, and 1.5-1.9\% along the coastline. They estimate that 500 kilotonnes of positively buoyant plastic enters the ocean each year, where 39-42\% originate from mismanaged waste along coastlines, 45-48\% originate from fishing-related activities (e.g. fishing lines, nets, traps, and crates), and 12-13\% from mismanaged waste entering the ocean via rivers.
@@ -40,7 +40,7 @@ Lagrangian ocean analysis, where virtual particles are tracked in hydrodynamic f
 
 The past decade has seen a growing number of community-developed software packages for performing Lagrangian simulations [@Paris2013; @Fredj2016; @Lange2017; @Doos2017; @Dagestad2018; @JalonRojas2019; @Delandmeter2019]. In many cases, these packages are specific to particular particle classes or hydrodynamic models, or are written and embedded in proprietary software languages, and can be inflexible or difficult to integrate into different applications. In the case of plastic dispersal simulations, the underlying physical processes are still being researched and their implementation is under development [@vanSebille2020]. Hence, an open-source, flexible, and modular approach to performing Lagrangian simulations is necessary for prototyping, developing, and testing new physical process parameterisation schemes. Easy-to-run simulations allow for a more reproducable results, and for simple-to-produce sensitivity analyses.
 
[…]
