### 2024-09-04 149e778
Update docs/paper/paper.md
Co-authored-by: Erik van Sebille <e.vansebille@uu.nl>
