### 2024-09-04 674390c
Adding 'in general' to dilineate the flexibility of parcels vs the flexibility of the kernels. Perhaps the first sentence should go to the prior paragraph, and we start this paragraph on this 'However, due to ...' sentence
