### 2024-09-04 eddc65b
[pre-commit.ci] auto fixes from pre-commit.com hooks
for more information, see https://pre-commit.ci
