### 2024-09-05 #7094 @michaeldenes
Hi @AnjaliSandip, we've now pushed the changes to the paper addressing the reviewer comments above, as well as adding the community guidelines and an API reference page to the readthedocs documentation, into the main branch of plasticparcels. Please let me know if there's anything else we need to do on our end. Thanks!
