### 2024-09-06 OceanParcels/plasticparcels#51 @philippemiron
examples
Hi all,

Would it be possible to include at least ~1-2 examples where all the input datasets are publicly available?

More precisely, after installing `plasticparcels`, I can't reproduce any of the examples, as e.g. [example_Italy_coast](https://plastic.oceanparcels.org/en/latest/examples/example_Italy_coast.html)  requires the model settings:

```
# Load the model settings
settings_file = 'docs/examples/example_Italy_coast_settings.json'
settings = pp.utils.load_settings(settings_file)
```
which are not part of the path.

Also, looking at those example configuration files [in the repo](https://github.com/OceanParcels/plasticparcels/tree/main/docs/examples), they all rely on private storage `/storage/shared/oceanparcels/input_data/`.

openjournals/joss-reviews#7094
