### 2024-09-12 OceanParcels/plasticparcels#51 @michaeldenes
Hi @philippemiron,

You've raised a good point, to run `plasticparcels` requires some hydrodynamic model data, biogeochemical model data, as well as wind and wave data. The example notebooks all use this model data, hence the storage filepath `/storage/shared/oceanparcels/input_data/`. We'll make this more clear in the `plasticparcels` documentation. We will, however, provide an example using an idealised fields that can be run independent of the required datasets.

In essence, `plasticparcels` can be run on any dataset that `parcels` can be run on, however we have configured it for Copernicus Marine Service data for now. Our near-future goal is to release `plasticparcels` on a cloud-platform, ideally in an environment with a variety of available hydrodynamic and biogeochemical datasets.

Expect to see an idealised example to try out soon!
