### 2024-09-12 OceanParcels/plasticparcels#51 @philippemiron
I think that will be very helpful to test out the kernels, and will also aligned a bit more with the rest of the parcel documentation.
