### 2024-09-13 71e362c
Starting progress on an idealised example tutorial.
