+++ docs/examples/example_idealised_flow.py
@@ -1,0 +1,2 @@
+def create_idealised_flow_fields():
+    return True
