### 2024-09-23 aaea287
Idealised flow example. Some text cleanup to do, but mostly there.
