+++ docs/examples/example_idealised_flow_functions.py
@@ -1,0 +1,244 @@
+import parcels
+import plasticparcels as pp
+
+from datetime import timedelta, datetime
+import numpy as np
+import xarray as xr
+
+
+def bickleyjet_fieldset_3d(times, xdim=51, ydim=51, zdim=11):
+    """ A Bickley Jet Field based on Hadjighasem et al 2017, 10.1063/1.4982720,
+    with a vertical dimension introduced via a linear decay of the 2D field in depth."""
+
+    U0 = 0.06266
+    L = 1770.0
+    r0 = 6371.0
+    k1 = 2 * 1 / r0
+    k2 = 2 * 2 / r0
+    k3 = 2 * 3 / r0
+    eps1 = 0.075
+    eps2 = 0.4
+    eps3 = 0.3
+    c3 = 0.461 * U0
+    c2 = 0.205 * U0
+    c1 = c3 + ((np.sqrt(5) - 1) / 2.0) * (k2 / k1) * (c2 - c3)
+
+    a, b = np.pi * r0, 7000.0  # domain size
+    lon = np.linspace(0, a, xdim, dtype=np.float32)
+    lat = np.linspace(-b / 2, b / 2, ydim, dtype=np.float32)
+    depth = np.linspace(0, 1000, zdim, dtype=np.float32)
+    dx, dy = lon[2] - lon[1], lat[2] - lat[1]
+
+    U = np.zeros((times.size, depth.size, lat.size, lon.size), dtype=np.float32)
+    V = np.zeros((times.size, depth.size, lat.size, lon.size), dtype=np.float32)
+    W = np.zeros((times.size, depth.size, lat.size, lon.size), dtype=np.float32)
+
+    for i in range(lon.size):
+        for j in range(lat.size):
+            x1 = lon[i] - dx / 2
+            x2 = lat[j] - dy / 2
+            for t in range(len(times)):
+                time = times[t]
+
+                f1 = eps1 * np.exp(-1j * k1 * c1 * time)
+                f2 = eps2 * np.exp(-1j * k2 * c2 * time)
+                f3 = eps3 * np.exp(-1j * k3 * c3 * time)
+                F1 = f1 * np.exp(1j * k1 * x1)
+                F2 = f2 * np.exp(1j * k2 * x1)
+                F3 = f3 * np.exp(1j * k3 * x1)
+                G = np.real(np.sum([F1, F2, F3]))
+                G_x = np.real(np.sum([1j * k1 * F1, 1j * k2 * F2, 1j * k3 * F3]))
+                U[t, 0, j, i] = (
+                    U0 / (np.cosh(x2 / L) ** 2)
+                    + 2 * U0 * np.sinh(x2 / L) / (np.cosh(x2 / L) ** 3) * G
+                )
+                V[t, 0, j, i] = U0 * L * (1.0 / np.cosh(x2 / L)) ** 2 * G_x
+    
+    # Add a linear decay in depth
+    for k in range(1, depth.size):
+        U[:, k, :, :] = U[:, 0, :, :] * ( (depth.size - k) / depth.size)
+        V[:, k, :, :] = V[:, 0, :, :] * ( (depth.size - k) / depth.size)
+
+    # Construct the fieldset
+    data = {"U": U, "V": V, "W": W}
+    dimensions = {"lon": lon, "lat": lat, "depth": depth, "time": times}
+    allow_time_extrapolation = True if len(times) == 1 else False
+    fieldset = parcels.FieldSet.from_data(
+        data, dimensions, mesh="flat", allow_time_extrapolation=allow_time_extrapolation
+    )
+
+    fieldset.U.interp_method = "cgrid_velocity"
+    fieldset.V.interp_method = "cgrid_velocity"
[…]
