+++ docs/examples/example_idealised_flow.py
@@ -1,2 +1,0 @@
-def create_idealised_flow_fields():
-    return True
