### 2024-09-23 #7094 @michaeldenes
Hi @AnjaliSandip, I'm currently finishing up an additional idealised flow tutorial and updating other documentation in the following branch https://github.com/OceanParcels/plasticparcels/tree/idealised_example, which I should have wrapped up by the end of the week. I'll update this issue once we pull-request those changes in to main for review!
