### 2024-09-24 f5ecf19
Updating the 'through surface' kernel to place the particle at the ocean surface (rather than delete)
