### 2024-09-24 71b3303
Updating the error through bathymetry kernel to use floats instead of integers.
