### 2024-09-24 2d0f9ea
Replacing integers with floats in the seawater density kernel
