### 2024-09-24 0afa892
Replacing integers with floats in the settling velocity kernel
