### 2024-09-24 f569f25
Replacing integers with floats in the biofouling kernel.
