### 2024-09-24 2201a72
Fix ruff issue
