### 2024-09-24 fb13a4b
Update the github readme to include a section on required data
