### 2024-09-24 b3daafa
Updating readme text re: required data.
