+++ README.md
@@ -24,9 +24,9 @@ conda install conda-forge::plasticparcels
 
 For the wind and wave data, we recommend using the [CDS API](https://cds.climate.copernicus.eu/api-how-to).
 
-To run the examples, you will need to update the data directories in settings `.json` files.
+To run most of the examples, you will need to update the data directories in settings `.json` files.
 
-Just like the `parcels` framework, `plasticparcels` can be adapted to use other hydrodynamic, biogeochemical, wave, and atmospheric models. If you require assistance, please contact us through the [Discussions page on GitHub](https://github.com/OceanParcels/plasticparcels/discussions).
+Just like the `parcels` framework, `plasticparcels` can be adapted to use other hydrodynamic, biogeochemical, wave, and atmospheric models. If you require assistance, please contact us through the [`plasticparcels` discussions page](https://github.com/OceanParcels/plasticparcels/discussions).
 
 ### Community contributions and support
 #### Contributing code
