### 2024-09-24 9b2d0b7
Updates to the idealised example to use 'from_modulefile' method rather than importing the functions and creating in script.
