### 2024-09-25 OceanParcels/plasticparcels#51 @michaeldenes
Hi @philippemiron, we've now created an idealised example that requires no additional data (see #55), and have updated the documentation for all the other examples. We'll roll these out in a new release soon, but they are all in the main branch now. Let me know if you have any issues, feel free to close this issue if all works! Cheers!
