### 2024-09-29 OceanParcels/plasticparcels#51 @philippemiron
I just reviewed the example and it is great, but I had a problem with the `conda-forge` package on two machines (intel osx and arm64 osx) running the first cell:

```
Cell In[1], line 7
      4 import numpy as np
      6 # parcels and plasticparcels imports
----> 7 import plasticparcels as pp
      8 import parcels
     10 # Plotting imports

File [~/micromamba/envs/test/lib/python3.12/site-packages/plasticparcels/__init__.py:5](http://localhost:8888/lab/tree/Downloads/plasticparcels/docs/examples/~/micromamba/envs/test/lib/python3.12/site-packages/plasticparcels/__init__.py#line=4)
      2 from plasticparcels.particlefile import *  # noqa
      3 from plasticparcels.utils import *  # noqa
----> 5 from plasticparcels._version_setup import __version__, __version_tuple__  # noqa

ModuleNotFoundError: No module named 'plasticparcels._version_setup'
```
It worked once I installed the current main branch of `plasticparcels` using `pip`... I will only recommend we use the next version of `plasticparcels` alongside the JOSS paper so the example is included.
