### 2024-09-29 OceanParcels/plasticparcels#51 @philippemiron
Small comment, in the last cell where you are comparing the 5 experiments, I think `ax.set_ylabel('Depth [m]')` should be `ax.set_ylabel('Meridional distance [m]')`.
