### 2024-09-29 #7094 @philippemiron
@AnjaliSandip, I have reviewed the example part of https://github.com/OceanParcels/plasticparcels/issues/51#issuecomment-2381096384 and this completes my review.

My only suggestion is to publish the JOSS paper using the next release of `plasticparcels`, so this new example is included in the artifact.
