+++ codecov.yml
@@ -1,0 +1,5 @@
+comment:
+  false
+
+# When modifying this file, please validate using
+# curl -X POST --data-binary @codecov.yml https://codecov.io/validate
