### 2024-09-30 3aeeabc
update linkcode resolve
