### 2024-09-30 OceanParcels/plasticparcels#51 @michaeldenes
Thanks for raising the versioning issue @philippemiron, this should be fixed by #59, and I've updated the y-axis labelling in #58. Cheers!
