### 2024-10-07 #7094 @AnjaliSandip
@michaeldenes Can you deposit a new archive of the software with Zenodo/figshare? After you have, please update this thread with archive DOI: `@editorialbot set 10.5281/zenodo.xxxxxx as archive`
