### 2024-10-17 #7094 @kyleniemeyer
@michaeldenes I'm not quite sure what is going on here. Can you start by removing all the `url` fields from the references? That could be an issue, though I'm not sure.

@openjournals/dev any ideas? Could this be related to any of our recent issues?
