### 2024-10-17 #7094 @philippemiron
To help debug I tried to run it locally.

```
git clone git@github.com:OceanParcels/plasticparcels.git
cd plasticparcels/docs
docker run --rm --volume $PWD/paper:/data --user $(id -u):$(id -g) --env JOURNAL=joss openjournals/inara
```

I get the same error:
```
INFO] Running filter /usr/local/share/openjournals/data/filters/parse-latex.lua
[INFO] Completed filter /usr/local/share/openjournals/data/filters/parse-latex.lua in 35 ms
[INFO] Running filter /usr/local/share/openjournals/data/filters/inline-cited-references.lua
[INFO] Loaded paper.bib from paper.bib
[INFO] Completed filter /usr/local/share/openjournals/data/filters/inline-cited-references.lua in 63 m
...
...
...
  \end{document}
[INFO] [makePDF] LaTeX run number 1
[INFO] [makePDF] LaTeX output
  Rc files read:
    NONE
  Latexmk: Run number 1 of rule 'lualatex'
  This is LuaHBTeX, Version 1.18.0 (TeX Live 2024)
   restricted system commands enabled.
  Latexmk: Getting log file '/tmp/tex2pdf.-1a7b73b90c72efab/input.log'
  Latexmk: Run number 2 of rule 'lualatex'
  This is LuaHBTeX, Version 1.18.0 (TeX Live 2024)
   restricted system commands enabled.
  Latexmk: Getting log file '/tmp/tex2pdf.-1a7b73b90c72efab/input.log'
```

but the PDF is output and looks fine...
```
> git status
On branch main
Your branch is up to date with 'origin/main'.

Untracked files:
  (use "git add <file>..." to include in what will be committed)
	paper/jats/
	paper/paper.pdf

nothing added to commit but untracked files present (use "git add" to track)
```
