### 2024-10-21 #7094 @michaeldenes
Hi @kyleniemeyer, happy to do so. Is there another way JOSS recommends to include hyperlinks?
