### 2024-10-21 #7094 @kyleniemeyer
> Hi @kyleniemeyer, happy to do so. Is there another way JOSS recommends to include hyperlinks?

Ah, with the DOI specified, there's no reason to also include the URL field - that's all I meant. However, we think the issue was the archive set here in the review issue.
