### 2024-10-22 #7094 @kthyng
@michaeldenes Can you change the author list and title in your Zenodo archive to exactly match the JOSS paper? This gives a consistent archive.
