### 2024-10-23 #7094 @michaeldenes
Thanks @kthyng, I've now made sure the zenodo archive has the same title and author list as the paper. Let me know if there's anything else missing. Cheers!
